Extract BillingClient state checks from MainActivity
