+++ app/src/main/kotlin/com/google/ai/sample/feature/multimodal/PhotoReasoningChatBubbles.kt
@@ -1,0 +1,191 @@
+package com.google.ai.sample.feature.multimodal
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.requiredSize
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Person
+import androidx.compose.material.icons.rounded.Refresh
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawBehind
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+
+@Composable
+fun StopButton(onClick: () -> Unit) {
+    Button(
+        onClick = onClick,
+        colors = ButtonDefaults.buttonColors(containerColor = Color.Red),
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(8.dp)
+    ) {
+        Text("Stop", color = Color.White)
+    }
+}
+
+@Composable
+fun UserChatBubble(
+    text: String,
+    isPending: Boolean,
+    imageUris: List<String> = emptyList(),
+    showUndo: Boolean = false,
+    onUndoClicked: () -> Unit = {}
+) {
+    Row(
+        modifier = Modifier
+            .padding(vertical = 8.dp, horizontal = 8.dp)
+            .fillMaxWidth(),
+        verticalAlignment = Alignment.Top
+    ) {
+        if (showUndo) {
+            IconButton(
+                onClick = onUndoClicked,
+                modifier = Modifier
+                    .padding(end = 8.dp)
+                    .align(Alignment.CenterVertically)
+            ) {
+                Icon(
+                    imageVector = Icons.Rounded.Refresh,
+                    contentDescription = "Undo",
+                    tint = Color.Gray
+                )
+            }
+        }
+        Spacer(modifier = Modifier.weight(1f))
+        Card(
+            shape = MaterialTheme.shapes.medium,
+            colors = CardDefaults.cardColors(
+                containerColor = MaterialTheme.colorScheme.primaryContainer
+            ),
+            modifier = Modifier.weight(4f)
+        ) {
+            Column(
+                modifier = Modifier.padding(all = 16.dp)
+            ) {
+                Text(
+                    text = text,
+                    color = MaterialTheme.colorScheme.onPrimaryContainer
+                )
+                if (imageUris.isNotEmpty()) {
+                    LazyRow(
+                        modifier = Modifier.padding(top = 8.dp)
+                    ) {
+                        items(imageUris) { uri ->
+                            AsyncImage(
+                                model = uri,
+                                contentDescription = null,
+                                modifier = Modifier
+                                    .padding(4.dp)
+                                    .requiredSize(100.dp)
+                            )
+                        }
+                    }
+                }
+                if (isPending) {
+                    CircularProgressIndicator(
+                        modifier = Modifier
+                            .padding(top = 8.dp)
+                            .requiredSize(16.dp),
+                        strokeWidth = 2.dp
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun ModelChatBubble(
+    text: String,
+    isPending: Boolean
+) {
+    Row(
+        modifier = Modifier
+            .padding(vertical = 8.dp, horizontal = 8.dp)
+            .fillMaxWidth(),
+        verticalAlignment = Alignment.Top
+    ) {
+        Card(
+            shape = MaterialTheme.shapes.medium,
+            colors = CardDefaults.cardColors(
+                containerColor = MaterialTheme.colorScheme.secondaryContainer
+            ),
+            modifier = Modifier.weight(4f)
+        ) {
+            Row(
+                modifier = Modifier.padding(all = 16.dp)
+            ) {
+                Icon(
+                    Icons.Outlined.Person,
+                    contentDescription = "AI Assistant",
+                    tint = MaterialTheme.colorScheme.onSecondaryContainer,
+                    modifier = Modifier
+                        .requiredSize(24.dp)
+                        .drawBehind {
+                            drawCircle(color = Color.White)
+                        }
+                        .padding(end = 8.dp)
+                )
+                Column {
+                    Text(
+                        text = text,
+                        color = MaterialTheme.colorScheme.onSecondaryContainer
+                    )
+                    if (isPending) {
+                        CircularProgressIndicator(
+                            modifier = Modifier
+                                .padding(top = 8.dp)
+                                .requiredSize(16.dp),
+                            strokeWidth = 2.dp
+                        )
+                    }
+                }
+            }
+        }
+        Spacer(modifier = Modifier.weight(1f))
+    }
+}
+
+@Composable
+fun ErrorChatBubble(
+    text: String
+) {
+    Box(
+        modifier = Modifier
+            .padding(vertical = 8.dp, horizontal = 8.dp)
+            .fillMaxWidth()
+    ) {
+        Card(
+            shape = MaterialTheme.shapes.medium,
+            colors = CardDefaults.cardColors(
+                containerColor = MaterialTheme.colorScheme.errorContainer
+            ),
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Text(
+                text = text,
+                color = MaterialTheme.colorScheme.error,
+                modifier = Modifier.padding(all = 16.dp)
+            )
+        }
+    }
+}
