+++ app/src/main/kotlin/com/google/ai/sample/feature/multimodal/PhotoReasoningCommandExecutionGuard.kt
@@ -1,0 +1,7 @@
+package com.google.ai.sample.feature.multimodal
+
+internal object PhotoReasoningCommandExecutionGuard {
+    fun shouldAbort(isJobActive: Boolean, isStopRequested: Boolean): Boolean {
+        return !isJobActive || isStopRequested
+    }
+}
