+++ app/src/main/kotlin/com/google/ai/sample/feature/multimodal/PhotoReasoningScreen.kt
@@ -122,133 +122,6 @@ import kotlinx.serialization.json.Json
 import android.util.Log
 import kotlinx.serialization.SerializationException
 
-// Define Colors
-val DarkYellow1 = Color(0xFFF0A500) // A darker yellow
-val DarkYellow2 = Color(0xFFF3C100) // A slightly lighter dark yellow
-
-@Composable
-fun StopButton(onClick: () -> Unit) {
-    Button(
-        onClick = onClick,
-        colors = ButtonDefaults.buttonColors(containerColor = Color.Red),
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp)
-    ) {
-        Text("Stop", color = Color.White)
-    }
-}
-
-@Composable
-internal fun PhotoReasoningRoute(
-    innerPadding: PaddingValues,  // Füge Parameter hinzu
-    viewModelStoreOwner: androidx.lifecycle.ViewModelStoreOwner = checkNotNull(
-        androidx.lifecycle.viewmodel.compose.LocalViewModelStoreOwner.current
-    ) { "ViewModelStoreOwner is required" }
-) {
-    val context = LocalContext.current
-    val mainActivity = context as? MainActivity
-    
-    // Scoped to MainActivity so it survives navigation and avoids duplicate initialization.
-    val owner = mainActivity ?: viewModelStoreOwner
-    val viewModel: PhotoReasoningViewModel = androidx.lifecycle.viewmodel.compose.viewModel(
-        viewModelStoreOwner = owner, 
-        factory = GenerativeViewModelFactory
-    )
-    val photoReasoningUiState by viewModel.uiState.collectAsState()
-    val commandExecutionStatus by viewModel.commandExecutionStatus.collectAsState()
-    val detectedCommands by viewModel.detectedCommands.collectAsState()
-    val systemMessage by viewModel.systemMessage.collectAsState()
-    val isInitialized by viewModel.isInitialized.collectAsState()
-    val modelName by viewModel.modelNameState.collectAsState()
-    val userInput by viewModel.userInput.collectAsState()
-    val isGenerationRunning by viewModel.isGenerationRunningFlow.collectAsState()
-    val isOfflineGpuModelLoaded by viewModel.isOfflineGpuModelLoadedFlow.collectAsState()
-    val isInitializingOfflineModel by viewModel.isInitializingOfflineModelFlow.collectAsState()
-
-    // Hoisted: var showNotificationRationaleDialog by rememberSaveable { mutableStateOf(false) }
-    // This state will now be managed in PhotoReasoningRoute and passed down.
-
-
-    val coroutineScope = rememberCoroutineScope()
-    val imageRequestBuilder = ImageRequest.Builder(LocalContext.current)
-    val imageLoader = ImageLoader.Builder(LocalContext.current).build()
-
-    val isAccessibilityServiceEffectivelyEnabled by mainActivity?.isAccessibilityServiceEnabledFlow?.collectAsState() ?: mutableStateOf(false)
-    val isMediaProjectionPermissionGranted by mainActivity?.isMediaProjectionPermissionGrantedFlow?.collectAsState() ?: mutableStateOf(false)
-    val isKeyboardOpen by mainActivity?.isKeyboardOpen?.collectAsState() ?: mutableStateOf(false)
-
-    val accessibilitySettingsLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.StartActivityForResult()
-    ) {
-        mainActivity?.refreshAccessibilityServiceStatus()
-    }
-
-    DisposableEffect(viewModel, mainActivity) {
-        mainActivity?.setPhotoReasoningViewModel(viewModel)
-        mainActivity?.refreshAccessibilityServiceStatus()
-        viewModel.loadSystemMessage(context)
-        viewModel.collectLiveApiMessages()
-        onDispose { }
-    }
-
-    PhotoReasoningScreen(
-        innerPadding = innerPadding,  // Übergebe an PhotoReasoningScreen
-        uiState = photoReasoningUiState,
-        commandExecutionStatus = commandExecutionStatus,
-        detectedCommands = detectedCommands,
-        systemMessage = systemMessage,
-        chatMessages = viewModel.chatMessagesFlow,
-        onSystemMessageChanged = { message ->
-            viewModel.updateSystemMessage(message, context)
-        },
-        onRestoreSystemMessageClicked = {
-            viewModel.restoreSystemMessage(context)
-        },
-        onReasonClicked = { inputText, selectedItems ->
-            coroutineScope.launch {
-                val bitmaps = selectedItems.mapNotNull {
-                    val imageRequest = imageRequestBuilder.data(it).precision(Precision.EXACT).build()
-                    try {
-                        val result = imageLoader.execute(imageRequest)
-                        if (result is SuccessResult) (result.drawable as BitmapDrawable).bitmap else null
-                    } catch (e: Exception) { null }
-                }
-                viewModel.reason(
-                    userInput = inputText,
-                    selectedImages = bitmaps,
-                    screenInfoForPrompt = null, // User-initiated messages don't have prior screen context here
-                    imageUrisForChat = selectedItems.map { it.toString() }
-                )
-            }
-        },
-        isAccessibilityServiceEnabled = isAccessibilityServiceEffectivelyEnabled,
-        isMediaProjectionPermissionGranted = isMediaProjectionPermissionGranted,
-        onEnableAccessibilityService = {
-            try {
-                // val intent = Intent(Settings.ACTION_ACCESSIBILITY_SETTINGS) // Corrected
-                // accessibilitySettingsLauncher.launch(intent) // Corrected below
-                val intent = Intent(Settings.ACTION_ACCESSIBILITY_SETTINGS)
-                accessibilitySettingsLauncher.launch(intent)
-            } catch (e: Exception) {
-                Toast.makeText(context, "Error opening Accessibility Settings.", Toast.LENGTH_LONG).show()
-            }
-        },
-        onClearChatHistory = {
-            mainActivity?.getPhotoReasoningViewModel()?.clearChatHistory(context)
-        },
-        isKeyboardOpen = isKeyboardOpen,
-        onStopClicked = { viewModel.onStopClicked() },
-        isInitialized = isInitialized,
-        modelName = modelName,
-        userQuestion = userInput,
-        onUserQuestionChanged = { viewModel.updateUserInput(it) },
-        isGenerationRunning = isGenerationRunning,
-        isOfflineGpuModelLoaded = isOfflineGpuModelLoaded,
-        isInitializingOfflineModel = isInitializingOfflineModel
-    )
-}
-
 @Composable
 fun PhotoReasoningScreen(
     innerPadding: PaddingValues,  // Füge Parameter hinzu
@@ -644,763 +517,3 @@ fun PhotoReasoningScreen(
             )
         }
     }
-
-
-@Composable
-fun DatabaseListPopup(
-    onDismissRequest: () -> Unit,
-    entries: List<SystemMessageEntry>,
-    onNewClicked: () -> Unit,
-    onEntryClicked: (SystemMessageEntry) -> Unit,
-    onDeleteClicked: (SystemMessageEntry) -> Unit,
-    onImportCompleted: () -> Unit
-) {
-    val TAG_IMPORT_PROCESS = "ImportProcess"
-    val scope = rememberCoroutineScope()
-    var entryMenuToShow: SystemMessageEntry? by remember { mutableStateOf(null) }
-    var selectionModeActive by rememberSaveable { mutableStateOf(false) }
-    var selectedEntryTitles by rememberSaveable { mutableStateOf(emptySet<String>()) }
-    var selectAllChecked by rememberSaveable { mutableStateOf(false) }
-    val context = LocalContext.current
-
-    var entryToConfirmOverwrite by remember { mutableStateOf<Pair<SystemMessageEntry, SystemMessageEntry>?>(null) }
-    var remainingEntriesToImport by remember { mutableStateOf<List<SystemMessageEntry>>(emptyList()) }
-    var skipAllDuplicates by remember { mutableStateOf(false) }
-
-    // processImportedEntries is defined within DatabaseListPopup, so it has access to context, onImportCompleted, etc.
-    fun processImportedEntries( 
-        imported: List<SystemMessageEntry>,
-        currentSystemEntries: List<SystemMessageEntry>
-    ) {
-        val TAG_IMPORT_PROCESS_FUNCTION = "ImportProcessFunction" 
-        Log.d(TAG_IMPORT_PROCESS_FUNCTION, "Starting processImportedEntries. Imported: ${imported.size}, Current: ${currentSystemEntries.size}, SkipAll: $skipAllDuplicates")
-
-        var newCount = 0
-        var updatedCount = 0 
-        var skippedCount = 0
-        val entriesToProcess = imported.toMutableList()
-
-        while (entriesToProcess.isNotEmpty()) {
-            val newEntry = entriesToProcess.removeAt(0)
-            Log.d(TAG_IMPORT_PROCESS_FUNCTION, "Processing entry: Title='${newEntry.title}'. Remaining in batch: ${entriesToProcess.size}")
-            val existingEntry = currentSystemEntries.find { it.title.equals(newEntry.title, ignoreCase = true) }
-
-            if (existingEntry != null) {
-                Log.d(TAG_IMPORT_PROCESS_FUNCTION, "Duplicate found for title: '${newEntry.title}'. Existing guide: '${existingEntry.guide.take(50)}', New guide: '${newEntry.guide.take(50)}'")
-                if (skipAllDuplicates) {
-                    Log.d(TAG_IMPORT_PROCESS_FUNCTION, "Skipping duplicate '${newEntry.title}' due to skipAllDuplicates flag.")
-                    skippedCount++
-                    continue
-                }
-                Log.d(TAG_IMPORT_PROCESS_FUNCTION, "Calling askForOverwrite for '${newEntry.title}'.")
-                entryToConfirmOverwrite = Pair(existingEntry, newEntry) 
-                remainingEntriesToImport = entriesToProcess.toList() 
-                return 
-            } else {
-                Log.i(TAG_IMPORT_PROCESS_FUNCTION, "Adding new entry: Title='${newEntry.title}'")
-                SystemMessageEntryPreferences.addEntry(context, newEntry)
-                newCount++
-            }
-        }
-        Log.i(TAG_IMPORT_PROCESS_FUNCTION, "Finished processing batch. newCount=$newCount, updatedCount=$updatedCount, skippedCount=$skippedCount")
-        val summary = "Import finished: $newCount added, $updatedCount updated, $skippedCount skipped."
-        Toast.makeText(context, summary as CharSequence, Toast.LENGTH_LONG).show()
-        onImportCompleted() 
-        skipAllDuplicates = false 
-    }
-
-
-    val filePickerLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.GetContent(),
-        onResult = { uri: Uri? ->
-            Log.d(TAG_IMPORT_PROCESS, "FilePickerLauncher onResult triggered.")
-            if (uri == null) {
-                Log.w(TAG_IMPORT_PROCESS, "URI is null, no file selected or operation cancelled.")
-                scope.launch(Dispatchers.Main) {
-                    Toast.makeText(context, "No file selected." as CharSequence, Toast.LENGTH_SHORT).show()
-                }
-                return@rememberLauncherForActivityResult
-            }
-
-            Log.i(TAG_IMPORT_PROCESS, "Selected file URI: $uri")
-            scope.launch(Dispatchers.Main) {
-                Toast.makeText(context, "File selected: $uri. Starting import..." as CharSequence, Toast.LENGTH_SHORT).show()
-            }
-
-            scope.launch(Dispatchers.IO) { 
-                try {
-                    Log.d(TAG_IMPORT_PROCESS, "Attempting to open InputStream for URI: $uri on thread: ${Thread.currentThread().name}")
-                    
-                    var fileSize = -1L
-                    try {
-                        context.contentResolver.openFileDescriptor(uri, "r")?.use { pfd ->
-                            fileSize = pfd.statSize
-                        }
-                        Log.i(TAG_IMPORT_PROCESS, "Estimated file size: $fileSize bytes.")
-                    } catch (e: Exception) {
-                        Log.w(TAG_IMPORT_PROCESS, "Could not determine file size for URI: $uri. Will proceed without size check.", e)
-                    }
-
-                    val MAX_FILE_SIZE_BYTES = 10 * 1024 * 1024 
-                    if (fileSize != -1L && fileSize > MAX_FILE_SIZE_BYTES) {
-                        Log.e(TAG_IMPORT_PROCESS, "File size ($fileSize bytes) exceeds limit of $MAX_FILE_SIZE_BYTES bytes.")
-                        withContext(Dispatchers.Main) {
-                            Toast.makeText(context, "File is too large (max 10MB)." as CharSequence, Toast.LENGTH_LONG).show()
-                        }
-                        return@launch 
-                    }
-                     if (fileSize == 0L) { 
-                         Log.w(TAG_IMPORT_PROCESS, "Imported file is empty (0 bytes).")
-                        withContext(Dispatchers.Main) {
-                            Toast.makeText(context, "Imported file is empty." as CharSequence, Toast.LENGTH_LONG).show()
-                        }
-                        return@launch 
-                    }
-
-                    context.contentResolver.openInputStream(uri)?.use { inputStream ->
-                        Log.i(TAG_IMPORT_PROCESS, "InputStream opened. Reading text on thread: ${Thread.currentThread().name}")
-                        val jsonString = inputStream.bufferedReader().readText() 
-                        Log.i(TAG_IMPORT_PROCESS, "File content read. Size: ${jsonString.length} chars.")
-                        Log.v(TAG_IMPORT_PROCESS, "File content snippet: ${jsonString.take(500)}")
-
-                        if (jsonString.isBlank()) {
-                            Log.w(TAG_IMPORT_PROCESS, "Imported file content is blank.")
-                            withContext(Dispatchers.Main) {
-                                Toast.makeText(context, "Imported file content is blank." as CharSequence, Toast.LENGTH_LONG).show()
-                            }
-                            return@use 
-                        }
-
-                        Log.d(TAG_IMPORT_PROCESS, "Attempting to parse JSON string on thread: ${Thread.currentThread().name}")
-                        val parsedEntries = Json.decodeFromString(ListSerializer(SystemMessageEntry.serializer()), jsonString)
-                        Log.i(TAG_IMPORT_PROCESS, "JSON parsed. Found ${parsedEntries.size} entries.")
-
-                        val currentSystemEntries = SystemMessageEntryPreferences.loadEntries(context) 
-                        Log.d(TAG_IMPORT_PROCESS, "Current system entries loaded: ${currentSystemEntries.size} entries.")
-
-                        withContext(Dispatchers.Main) { 
-                            Log.d(TAG_IMPORT_PROCESS, "Switching to Main thread for processImportedEntries: ${Thread.currentThread().name}")
-                            skipAllDuplicates = false 
-                            processImportedEntries(
-                                imported = parsedEntries,
-                                currentSystemEntries = currentSystemEntries
-                            )
-                        }
-                    } ?: Log.w(TAG_IMPORT_PROCESS, "ContentResolver.openInputStream returned null for URI: $uri (second check).")
-                } catch (oom: OutOfMemoryError) {
-                    Log.e(TAG_IMPORT_PROCESS, "Out of memory during file import for URI: $uri on thread: ${Thread.currentThread().name}", oom)
-                    withContext(Dispatchers.Main) {
-                        Toast.makeText(
-                            context,
-                            "Error importing file: Out of memory. File may be too large or contain too many entries." as CharSequence,
-                            Toast.LENGTH_LONG
-                        ).show()
-                    }
-                } catch (e: Exception) {
-                    Log.e(TAG_IMPORT_PROCESS, "Error during file import for URI: $uri on thread: ${Thread.currentThread().name}", e)
-                    withContext(Dispatchers.Main) {
-                        val errorMessage = e.message ?: "Unknown error during import."
-                        Toast.makeText(context, "Error importing file: $errorMessage" as CharSequence, Toast.LENGTH_LONG).show()
-                    }
-                }
-            }
-        }
-    )
-
-    if (entryToConfirmOverwrite != null) {
-        val (existingEntry, newEntry) = checkNotNull(entryToConfirmOverwrite)
-        OverwriteConfirmationDialog(
-            entryTitle = newEntry.title,
-            onConfirm = {
-                Log.d(TAG_IMPORT_PROCESS, "Overwrite confirmed for title: '${newEntry.title}'")
-                SystemMessageEntryPreferences.updateEntry(context, existingEntry, newEntry)
-                Toast.makeText(context, "Entry '${newEntry.title}' overwritten." as CharSequence, Toast.LENGTH_SHORT).show()
-                entryToConfirmOverwrite = null
-                val currentSystemEntriesAfterUpdate = SystemMessageEntryPreferences.loadEntries(context) 
-                Log.d(TAG_IMPORT_PROCESS, "Continuing with remaining ${remainingEntriesToImport.size} entries after dialog (Confirm).")
-                processImportedEntries( 
-                    imported = remainingEntriesToImport,
-                    currentSystemEntries = currentSystemEntriesAfterUpdate
-                )
-            },
-            onDeny = { 
-                Log.d(TAG_IMPORT_PROCESS, "Overwrite denied for title: '${newEntry.title}'")
-                entryToConfirmOverwrite = null
-                val currentSystemEntriesAfterDeny = SystemMessageEntryPreferences.loadEntries(context) 
-                Log.d(TAG_IMPORT_PROCESS, "Continuing with remaining ${remainingEntriesToImport.size} entries after dialog (Deny).")
-                processImportedEntries( 
-                    imported = remainingEntriesToImport,
-                    currentSystemEntries = currentSystemEntriesAfterDeny
-                )
-            },
-            onSkipAll = { 
-                Log.d(TAG_IMPORT_PROCESS, "Skip All selected for title: '${newEntry.title}'")
-                skipAllDuplicates = true
-                entryToConfirmOverwrite = null
-                val currentSystemEntriesAfterSkipAll = SystemMessageEntryPreferences.loadEntries(context) 
-                Log.d(TAG_IMPORT_PROCESS, "Continuing with remaining ${remainingEntriesToImport.size} entries after dialog (SkipAll).")
-                processImportedEntries( 
-                    imported = remainingEntriesToImport,
-                    currentSystemEntries = currentSystemEntriesAfterSkipAll
-                )
-            },
-            onDismiss = { 
-                Log.d(TAG_IMPORT_PROCESS, "Overwrite dialog dismissed for title: '${entryToConfirmOverwrite?.second?.title}'. Import process for this batch might halt.")
-                entryToConfirmOverwrite = null 
-                remainingEntriesToImport = emptyList() 
-                skipAllDuplicates = false
-                Toast.makeText(context, "Import cancelled for remaining items." as CharSequence, Toast.LENGTH_SHORT).show()
-                onImportCompleted() 
-            }
-        )
-    }
-
-
-    Dialog(
-        onDismissRequest = onDismissRequest,
-        properties = DialogProperties(usePlatformDefaultWidth = false)
-    ) {
-        Card(
-            modifier = Modifier
-                .fillMaxWidth(0.95f) 
-                .fillMaxHeight(0.85f), 
-            shape = RoundedCornerShape(16.dp),
-            colors = CardDefaults.cardColors(containerColor = DarkYellow1)
-        ) {
-            Column(
-                modifier = Modifier
-                    .padding(16.dp) 
-                    .fillMaxSize()
-            ) {
-                val displayRowCount = 15
-                val newButtonRowIndex = entries.size 
-
-                LazyColumn(modifier = Modifier.weight(1f)) { 
-                    items(displayRowCount) { rowIndex ->
-                        val currentAlternatingColor = if (rowIndex % 2 == 0) DarkYellow1 else DarkYellow2
-
-                        when {
-                            rowIndex < entries.size -> {
-                                val entry = entries[rowIndex]
-                                val isSelected = selectedEntryTitles.contains(entry.title)
-
-                                Row(
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .background(currentAlternatingColor)
-                                        .clickable {
-                                            if (selectionModeActive) {
-                                                val entryTitle = entry.title
-                                                val isCurrentlySelected = selectedEntryTitles.contains(entryTitle)
-                                                selectedEntryTitles = if (isCurrentlySelected) {
-                                                    selectedEntryTitles - entryTitle
-                                                } else {
-                                                    selectedEntryTitles + entryTitle
-                                                }
-                                                selectAllChecked = if (selectedEntryTitles.size == entries.size && entries.isNotEmpty()) true else if (selectedEntryTitles.isEmpty()) false else false
-                                            } else {
-                                                onEntryClicked(entry) 
-                                            }
-                                        }
-                                        .padding(horizontal = 8.dp, vertical = 4.dp),
-                                    verticalAlignment = Alignment.CenterVertically
-                                ) {
-                                    if (selectionModeActive) {
-                                        Checkbox(
-                                            checked = isSelected,
-                                            onCheckedChange = { isChecked ->
-                                                val entryTitle = entry.title
-                                                selectedEntryTitles = if (isChecked) {
-                                                    selectedEntryTitles + entryTitle
-                                                } else {
-                                                    selectedEntryTitles - entryTitle
-                                                }
-                                                selectAllChecked = if (selectedEntryTitles.size == entries.size && entries.isNotEmpty()) true else if (selectedEntryTitles.isEmpty()) false else false
-                                            },
-                                            modifier = Modifier.padding(end = 8.dp),
-                                            colors = CheckboxDefaults.colors(checkedColor = MaterialTheme.colorScheme.primary)
-                                        )
-                                    }
-                                    Box(
-                                        modifier = Modifier
-                                            .weight(1f)
-                                            .border(BorderStroke(1.dp, Color.Black), shape = RoundedCornerShape(12.dp))
-                                            .padding(horizontal = 12.dp, vertical = 8.dp),
-                                        contentAlignment = Alignment.CenterStart
-                                    ) {
-                                        Text(entry.title, color = Color.Black, style = MaterialTheme.typography.bodyLarge)
-                                    }
-                                    if (!selectionModeActive) {
-                                        Box {
-                                            IconButton(onClick = { entryMenuToShow = entry }) {
-                                                Icon(Icons.Filled.MoreVert, "More options", tint = Color.Black)
-                                            }
-                                            DropdownMenu(
-                                                expanded = entryMenuToShow == entry,
-                                                onDismissRequest = { entryMenuToShow = null }
-                                            ) {
-                                                DropdownMenuItem(
-                                                    text = { Text("Delete") },
-                                                    onClick = {
-                                                        onDeleteClicked(entry)
-                                                        entryMenuToShow = null
-                                                    }
-                                                )
-                                            }
-                                        }
-                                    } else {
-                                        Spacer(modifier = Modifier.width(48.dp)) 
-                                    }
-                                }
-                            }
-                            rowIndex == newButtonRowIndex -> {
-                                Row(
-                                    modifier = Modifier.fillMaxWidth().background(currentAlternatingColor)
-                                        .padding(8.dp).clickable(onClick = onNewClicked),
-                                    verticalAlignment = Alignment.CenterVertically,
-                                    horizontalArrangement = Arrangement.End 
-                                ) {
-                                    Text("This is also sent to the AI", color = Color.Black.copy(alpha = 0.6f), style = MaterialTheme.typography.bodyMedium, modifier = Modifier.weight(1f))
-                                    Button(onClick = onNewClicked, colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.primary), modifier = Modifier.padding(start = 8.dp)) {
-                                        Text("New")
-                                    }
-                                }
-                            }
-                            else -> {
-                                Box(modifier = Modifier.fillMaxWidth().height(56.dp).background(currentAlternatingColor).padding(16.dp)) {}
-                            }
-                        }
-                    }
-                }
-                Spacer(modifier = Modifier.height(8.dp))
-                Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween, verticalAlignment = Alignment.CenterVertically) {
-                    if (selectionModeActive) {
-                        Row(verticalAlignment = Alignment.CenterVertically) {
-                            Checkbox(
-                                checked = selectAllChecked,
-                                onCheckedChange = { isChecked ->
-                                    selectAllChecked = isChecked
-                                    selectedEntryTitles = if (isChecked) entries.map { it.title }.toSet() else emptySet()
-                                },
-                                colors = CheckboxDefaults.colors(checkedColor = MaterialTheme.colorScheme.primary)
-                            )
-                            Text("All", color = Color.Black, style = MaterialTheme.typography.bodyMedium)
-                        }
-                    } else {
-                        Spacer(modifier = Modifier.width(80.dp)) // Placeholder for alignment
-                    }
-
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Button(onClick = { filePickerLauncher.launch("*/*") }, colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.primary), modifier = Modifier.padding(end = 8.dp)) { Text("Import") }
-                        Button(
-                            onClick = {
-                                if (selectionModeActive) { 
-                                    if (selectedEntryTitles.isEmpty()) {
-                                        Toast.makeText(context, "No entries selected for export." as CharSequence, Toast.LENGTH_SHORT).show()
-                                    } else {
-                                        val entriesToExport = entries.filter { selectedEntryTitles.contains(it.title) }
-                                        val jsonString = Json.encodeToString(ListSerializer(SystemMessageEntry.serializer()), entriesToExport)
-                                        shareTextFile(context, "system_messages_export.txt", jsonString)
-                                    }
-                                    selectionModeActive = false
-                                    selectedEntryTitles = emptySet()
-                                    selectAllChecked = false
-                                } else { 
-                                    selectionModeActive = true
-                                }
-                            },
-                            colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.primary)
-                        ) { Text("Export") } // Text is now always "Export"
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-@Composable
-fun OverwriteConfirmationDialog(
-    entryTitle: String,
-    onConfirm: () -> Unit,
-    onDeny: () -> Unit,
-    onSkipAll: () -> Unit,
-    onDismiss: () -> Unit
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Confirm Overwrite") },
-        text = { Text("An entry with the title \"$entryTitle\" already exists. Do you want to overwrite its guide?") },
-        confirmButton = {
-            TextButton(onClick = onConfirm) { Text("Yes") }
-        },
-        dismissButton = {
-            Row {
-                TextButton(onClick = onSkipAll) { Text("Skip All") }
-                TextButton(onClick = onDeny) { Text("No") }
-            }
-        }
-    )
-}
-
-
-@Composable
-fun UserChatBubble(
-    text: String,
-    isPending: Boolean,
-    imageUris: List<String> = emptyList(),
-    showUndo: Boolean = false,
-    onUndoClicked: () -> Unit = {}
-) {
-    Row(
-        modifier = Modifier
-            .padding(vertical = 8.dp, horizontal = 8.dp)
-            .fillMaxWidth(),
-        verticalAlignment = Alignment.Top
-    ) {
-        if (showUndo) {
-            IconButton(
-                onClick = onUndoClicked,
-                modifier = Modifier
-                    .padding(end = 8.dp)
-                    .align(Alignment.CenterVertically)
-            ) {
-                Icon(
-                    imageVector = Icons.Rounded.Refresh,
-                    contentDescription = "Undo",
-                    tint = Color.Gray
-                )
-            }
-        }
-        Spacer(modifier = Modifier.weight(1f))
-        Card(
-            shape = MaterialTheme.shapes.medium,
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.primaryContainer
-            ),
-            modifier = Modifier.weight(4f)
-        ) {
-            Column(
-                modifier = Modifier.padding(all = 16.dp)
-            ) {
-                Text(
-                    text = text,
-                    color = MaterialTheme.colorScheme.onPrimaryContainer
-                )
-                if (imageUris.isNotEmpty()) {
-                    LazyRow(
-                        modifier = Modifier.padding(top = 8.dp)
-                    ) {
-                        items(imageUris) { uri ->
-                            AsyncImage(
-                                model = uri,
-                                contentDescription = null,
-                                modifier = Modifier
-                                    .padding(4.dp)
-                                    .requiredSize(100.dp)
-                            )
-                        }
-                    }
-                }
-                if (isPending) {
-                    CircularProgressIndicator(
-                        modifier = Modifier
-                            .padding(top = 8.dp)
-                            .requiredSize(16.dp),
-                        strokeWidth = 2.dp
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun ModelChatBubble(
-    text: String,
-    isPending: Boolean
-) {
-    Row(
-        modifier = Modifier
-            .padding(vertical = 8.dp, horizontal = 8.dp)
-            .fillMaxWidth(),
-        verticalAlignment = Alignment.Top
-    ) {
-        Card(
-            shape = MaterialTheme.shapes.medium,
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.secondaryContainer
-            ),
-            modifier = Modifier.weight(4f)
-        ) {
-            Row(
-                modifier = Modifier.padding(all = 16.dp)
-            ) {
-                Icon(
-                    Icons.Outlined.Person,
-                    contentDescription = "AI Assistant",
-                    tint = MaterialTheme.colorScheme.onSecondaryContainer,
-                    modifier = Modifier
-                        .requiredSize(24.dp)
-                        .drawBehind {
-                            drawCircle(color = Color.White)
-                        }
-                        .padding(end = 8.dp)
-                )
-                Column {
-                    Text(
-                        text = text,
-                        color = MaterialTheme.colorScheme.onSecondaryContainer
-                    )
-                    if (isPending) {
-                        CircularProgressIndicator(
-                            modifier = Modifier
-                                .padding(top = 8.dp)
-                                .requiredSize(16.dp),
-                            strokeWidth = 2.dp
-                        )
-                    }
-                }
-            }
-        }
-        Spacer(modifier = Modifier.weight(1f))
-    }
-}
-
-@Composable
-fun ErrorChatBubble(
-    text: String
-) {
-    Box(
-        modifier = Modifier
-            .padding(vertical = 8.dp, horizontal = 8.dp)
-            .fillMaxWidth()
-    ) {
-        Card(
-            shape = MaterialTheme.shapes.medium,
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.errorContainer
-            ),
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text(
-                text = text,
-                color = MaterialTheme.colorScheme.error,
-                modifier = Modifier.padding(all = 16.dp)
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-fun PhotoReasoningScreenPreviewWithContent() {
-    MaterialTheme {
-        PhotoReasoningScreen(
-            innerPadding = PaddingValues(),
-            uiState = PhotoReasoningUiState.Success("This is a preview of the photo reasoning screen."),
-            commandExecutionStatus = "Command executed: Take screenshot",
-            detectedCommands = listOf(
-                Command.TakeScreenshot,
-                Command.ClickButton("OK")
-            ),
-            systemMessage = "This is a system message for the AI",
-            chatMessages = MutableStateFlow(listOf(
-                PhotoReasoningMessage(text = "Hello, how can I help you?", participant = PhotoParticipant.USER),
-                PhotoReasoningMessage(text = "I am here to help you. What do you want to know?", participant = PhotoParticipant.MODEL)
-            )),
-            isKeyboardOpen = false,
-            onStopClicked = {},
-            isInitialized = true
-        )
-    }
-}
-
-@Composable
-fun EditEntryPopup(
-    entry: SystemMessageEntry?,
-    onDismissRequest: () -> Unit,
-    onSaveClicked: (title: String, guide: String, originalEntry: SystemMessageEntry?) -> Unit
-) {
-    Dialog(
-        onDismissRequest = onDismissRequest,
-        properties = DialogProperties(usePlatformDefaultWidth = false) 
-    ) {
-        Card(
-            modifier = Modifier.fillMaxWidth(0.9f).fillMaxHeight(0.7f).padding(16.dp), 
-            shape = RoundedCornerShape(16.dp),
-            colors = CardDefaults.cardColors(containerColor = DarkYellow1) 
-        ) {
-            Column(modifier = Modifier.padding(16.dp).fillMaxSize()) {
-                var titleInput by rememberSaveable { mutableStateOf(entry?.title ?: "") }
-                var guideInput by rememberSaveable { mutableStateOf(entry?.guide ?: "") }
-
-                Text("Title", style = MaterialTheme.typography.labelMedium, color = Color.Black.copy(alpha = 0.7f)) 
-                Spacer(modifier = Modifier.height(4.dp))
-                OutlinedTextField(
-                    value = titleInput,
-                    onValueChange = { titleInput = it },
-                    placeholder = { Text("App/Task", color = Color.Gray) },
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = TextFieldDefaults.colors( 
-                        focusedContainerColor = Color.White,
-                        unfocusedContainerColor = Color.White,
-                        disabledContainerColor = Color.White,
-                        cursorColor = Color.Black, 
-                        focusedTextColor = Color.Black,
-                        unfocusedTextColor = Color.Black,
-                    ),
-                    singleLine = true
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                Text("Guide", style = MaterialTheme.typography.labelMedium, color = Color.Black.copy(alpha = 0.7f)) 
-                Spacer(modifier = Modifier.height(4.dp))
-                OutlinedTextField(
-                    value = guideInput,
-                    onValueChange = { guideInput = it },
-                    placeholder = { Text("Write a guide for an LLM on how it should perform certain tasks to be successful", color = Color.Gray) },
-                    modifier = Modifier.fillMaxWidth().weight(1f),
-                    colors = TextFieldDefaults.colors(
-                        focusedContainerColor = Color.White,
-                        unfocusedContainerColor = Color.White,
-                        disabledContainerColor = Color.White,
-                        cursorColor = Color.Black, 
-                        focusedTextColor = Color.Black,
-                        unfocusedTextColor = Color.Black,
-                    ),
-                    minLines = 5
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                Button(
-                    onClick = { onSaveClicked(titleInput, guideInput, entry) },
-                    modifier = Modifier.align(Alignment.End),
-                    colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.primary)
-                ) { Text("Save") }
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true, name = "EditEntryPopup New")
-@Composable
-fun EditEntryPopupNewPreview() {
-    MaterialTheme {
-        EditEntryPopup(entry = null, onDismissRequest = {}, onSaveClicked = { _, _, _ -> })
-    }
-}
-
-@Preview(showBackground = true, name = "EditEntryPopup Edit")
-@Composable
-fun EditEntryPopupEditPreview() {
-    MaterialTheme {
-        EditEntryPopup(entry = SystemMessageEntry("Existing Title", "Existing Guide"), onDismissRequest = {}, onSaveClicked = { _, _, _ -> })
-    }
-}
-
-val SystemMessageEntrySaver = Saver<SystemMessageEntry?, List<String?>>(
-    save = { entry -> if (entry == null) listOf(null, null) else listOf(entry.title, entry.guide) },
-    restore = { list ->
-        val title = list[0]
-        val guide = list[1]
-        if (title != null && guide != null) SystemMessageEntry(title, guide) else null
-    }
-)
-
-@Composable
-@Preview(showSystemUi = true)
-fun PhotoReasoningScreenPreviewEmpty() {
-    MaterialTheme {
-        PhotoReasoningScreen(
-            innerPadding = PaddingValues(),
-            chatMessages = MutableStateFlow<List<PhotoReasoningMessage>>(emptyList()),
-            isKeyboardOpen = false,
-            onStopClicked = {},
-            isInitialized = true
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DatabaseListPopupPreview() {
-    MaterialTheme {
-        DatabaseListPopup(
-            onDismissRequest = {},
-            entries = listOf(
-                SystemMessageEntry("Title 1", "Guide for prompt 1"),
-                SystemMessageEntry("Title 2", "Another guide for prompt 2"),
-                SystemMessageEntry("Title 3", "Yet another guide for prompt 3.")
-            ),
-            onNewClicked = {},
-            onEntryClicked = {},
-            onDeleteClicked = {},
-            onImportCompleted = {}
-        )
-    }
-}
-
-@Preview(showBackground = true, name = "DatabaseListPopup Empty")
-@Composable
-fun DatabaseListPopupEmptyPreview() {
-    MaterialTheme {
-        DatabaseListPopup(onDismissRequest = {}, entries = emptyList(), onNewClicked = {}, onEntryClicked = {}, onDeleteClicked = {}, onImportCompleted = {})
-    }
-}
-
-@Preview(showBackground = true, name = "Stop Button Preview")
-@Composable
-fun StopButtonPreview() {
-    MaterialTheme {
-        StopButton {}
-    }
-}
-
-@Composable
-fun VerticalScrollbar(
-    modifier: Modifier = Modifier,
-    listState: LazyListState
-) {
-    val scrollbarState by remember {
-        derivedStateOf {
-            val layoutInfo = listState.layoutInfo
-            val totalItems = layoutInfo.totalItemsCount
-            if (totalItems == 0) return@derivedStateOf null
-
-            val viewportHeight = layoutInfo.viewportSize.height.toFloat()
-            val firstVisibleItemIndex = listState.firstVisibleItemIndex
-            val visibleItemCount = layoutInfo.visibleItemsInfo.size
-
-            if (visibleItemCount >= totalItems) return@derivedStateOf null // All items visible, no scrollbar
-
-            val thumbHeight = (visibleItemCount.toFloat() / totalItems * viewportHeight)
-                .coerceAtLeast(20f)
-                .coerceAtMost(viewportHeight)
-
-            val maxScrollOffset = (viewportHeight - thumbHeight).coerceAtLeast(0f)
-            if (maxScrollOffset == 0f) return@derivedStateOf null
-
-            val firstItemOffset = if (layoutInfo.visibleItemsInfo.isNotEmpty()) {
-                val firstItem = layoutInfo.visibleItemsInfo.first()
-                if (firstItem.size > 0) listState.firstVisibleItemScrollOffset.toFloat() / firstItem.size else 0f
-            } else 0f
-
-            val scrollProgress = (firstVisibleItemIndex + firstItemOffset) /
-                (totalItems - visibleItemCount).coerceAtLeast(1)
-            val scrollOffset = (scrollProgress * maxScrollOffset).coerceIn(0f, maxScrollOffset)
-
-            Pair(scrollOffset, thumbHeight)
-        }
-    }
-
-    if (scrollbarState != null) {
-        val (offset, height) = checkNotNull(scrollbarState)
-         Canvas(modifier = modifier.width(4.dp)) {
-            drawRoundRect(
-                color = Color.Gray,
-                topLeft = Offset(0f, offset),
-                size = Size(size.width, height),
-                cornerRadius = CornerRadius(4.dp.toPx())
-            )
-        }
-    }
-}
