+++ app/src/main/kotlin/com/google/ai/sample/feature/multimodal/PhotoReasoningUiConstants.kt
@@ -1,0 +1,8 @@
+package com.google.ai.sample.feature.multimodal
+
+import androidx.compose.ui.graphics.Color
+
+// Define Colors
+val DarkYellow1 = Color(0xFFF0A500) // A darker yellow
+val DarkYellow2 = Color(0xFFF3C100) // A slightly lighter dark yellow
+
